requirements.txt
python-dotenv>=1.0.1
fandanGO-aria
logs-py